implement shared memory control interface for multi platform support (#743)
* implement shared memory control interface for multi platform support

unitracei-1

unitracei-1 implement shared memory linux fix

* Update shared_memory.h

Removed unused header

* Remove inapplicable comment

* Add comments

---------

+++ tools/unitrace/src/opencl/cl_collector.h
@@ -1347,8 +1347,8 @@ class ClCollector {
       collector->CalculateKernelGlobalSize(params, &instance->props);
       collector->CalculateKernelLocalSize(params, &instance->props);
 
-      uint64_t base_addr = 0;
-      uint64_t size = 0;
+      size_t base_addr = 0;
+      size_t size = 0;
 
       status = clGetKernelInfo(kernel, CL_KERNEL_BINARY_GPU_ADDRESS_INTEL, 0, nullptr, &size);
       status = clGetKernelInfo(kernel, CL_KERNEL_BINARY_GPU_ADDRESS_INTEL, size, &base_addr, &size);
